+++ src/azure-cli/azure/cli/command_modules/appservice/_constants.py
@@ -39,39 +39,18 @@ RUNTIME_TO_DEFAULT_VERSION_FUNCTIONAPP = {
         'java': '8'
     }
 }
-# functions version -> runtime -> runtime version : container image
-RUNTIME_TO_IMAGE_FUNCTIONAPP = {
+# functions version -> runtime : runtime versions
+RUNTIME_TO_SUPPORTED_VERSIONS_FUNCTIONAPP = {
     '2': {
-        'node': {
-            '8': 'mcr.microsoft.com/azure-functions/node:2.0-node8-appservice',
-            '10': 'mcr.microsoft.com/azure-functions/node:2.0-node10-appservice'
-        },
-        'python': {
-            '3.6': 'mcr.microsoft.com/azure-functions/python:2.0-python3.6-appservice',
-            '3.7': 'mcr.microsoft.com/azure-functions/python:2.0-python3.7-appservice'
-        },
-        'dotnet': {
-            '2': 'mcr.microsoft.com/azure-functions/dotnet:2.0-appservice'
-        },
-        'java': {
-            '8': 'mcr.microsoft.com/azure-functions/java:2.0-java8-appservice'
-        }
+        'node': ['8', '10'],
+        'python': ['3.6', '3.7'],
+        'dotnet': ['2'],
+        'java': ['8']
     },
     '3': {
-        'node': {
-            '10': 'mcr.microsoft.com/azure-functions/node:3.0-node10-appservice',
-            '12': 'mcr.microsoft.com/azure-functions/node:3.0-node12-appservice'
-        },
-        'python': {
-            '3.6': 'mcr.microsoft.com/azure-functions/python:3.0-python3.6-appservice',
-            '3.7': 'mcr.microsoft.com/azure-functions/python:3.0-python3.7-appservice',
-            '3.8': 'mcr.microsoft.com/azure-functions/python:3.0-python3.8-appservice'
-        },
-        'dotnet': {
-            '3': 'mcr.microsoft.com/azure-functions/dotnet:3.0-appservice'
-        },
-        'java': {
-            '8': 'mcr.microsoft.com/azure-functions/java:3.0-java8-appservice'
-        }
+        'node': ['10', '12'],
+        'python': ['3.6', '3.7', '3.8'],
+        'dotnet': ['3'],
+        'java': ['8']
     }
 }